[change] Improved logic for determining crossed threshold
