+++ openwisp_monitoring/db/backends/influxdb/client.py
@@ -131,7 +131,7 @@ class DatabaseClient(object):
     def use_udp(self):
         return TIMESERIES_DB.get("OPTIONS", {}).get("udp_writes", False)
 
-    def _get_operator(self, op):
+    def _clean_operator(self, op):
         """Returns the operator if it is valid."""
         if op not in self._OPERATORS:
             raise self.client_error(
@@ -139,6 +139,11 @@ class DatabaseClient(object):
                 f"Valid operators are: {', '.join(self._OPERATORS)}"
             )
         return op
+
+    def _clean_value(self, value):
+        if isinstance(value, datetime):
+            return f"'{self._get_timestamp(value)}'"
+        return value
 
     @retry
     def create_or_alter_retention_policy(self, name, duration):
@@ -263,6 +268,7 @@ class DatabaseClient(object):
         count_fields = kwargs.get("count_fields", [])
         rp = kwargs.get("retention_policy")
         where = kwargs.get("where", [])
+        precision = kwargs.get("precision", "s")
 
         # Ensure fields is a list (in case it's passed as a string)
         if isinstance(fields, str):
@@ -311,7 +317,8 @@ class DatabaseClient(object):
 
         conditions = []
         for field, op, value in where:
-            op = self._get_operator(op)
+            op = self._clean_operator(op)
+            value = self._clean_value(value)
             conditions.append(f"{field} {op} {value}")
         # Add conditions (time and tags)
         if since:
@@ -338,7 +345,7 @@ class DatabaseClient(object):
             q = f"{q} ORDER BY {order}"
         if limit:
             q = f"{q} LIMIT {limit}"
-        return list(self.query(q, precision="s").get_points())
+        return list(self.query(q, precision=precision).get_points())
 
     def get_list_query(self, query, precision="s"):
         result = self.query(query, precision=precision)
